requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.7.6
PyNaCl>=1.5.0
aiohttp>=3.8.0
